# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UpdateSortOrderBody": {
				"description": "更新排序号请求体",
				"properties": {
					"sort_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"sort_order"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CreateModelBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"model_key": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"aspect_ratio": {
						"type": "string"
					},
					"aspect_ratios": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"consume_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"consume_points"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateModelBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"model_key": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"aspect_ratio": {
						"type": "string"
					},
					"aspect_ratios": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"consume_points": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
			"AdminCreateCategoryBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"sort_order": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AdminUpdateCategoryBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"sort_order": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"AdminUpdateSortOrderBody": {
				"properties": {
					"sort_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"sort_order"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "普通用户认证 Token"
			},
			"adminJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "管理员认证 Token"
			}
		}
	},
	"info": {
		"title": "my-project",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
...
		"/app/categories/{id}/sort": {
			"put": {
				"operationId": "UpdateSortOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "更新分类排序号\nPUT /app/categories/:id/sort",
				"tags": [
					"分类管理"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSortOrderBody"
							}
						}
					}
				}
			}
		},
...
		"/app/models": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取模型列表\nGET /app/models",
				"tags": [
					"模型"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "manufacturer",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/admin/models": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateModelBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/admin/models/{id}": {
			"get": {
				"operationId": "Detail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateModelBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"模型管理"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/admin/categories": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取分类列表（分页）\nGET /api/v1/admin/categories",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "创建分类\nPOST /api/v1/admin/categories",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminCreateCategoryBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/admin/categories/enabled": {
			"get": {
				"operationId": "GetAllEnabled",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取所有启用分类\nGET /api/v1/admin/categories/enabled",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/admin/categories/{id}": {
			"get": {
				"operationId": "Detail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "获取分类详情\nGET /api/v1/admin/categories/:id",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "更新分类\nPUT /api/v1/admin/categories/:id",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminUpdateCategoryBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "删除分类\nDELETE /api/v1/admin/categories/:id",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/admin/categories/{id}/sort": {
			"put": {
				"operationId": "UpdateSortOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "更新分类排序号\nPUT /api/v1/admin/categories/:id/sort",
				"tags": [
					"管理端分类"
				],
				"security": [
					{
						"adminJwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminUpdateSortOrderBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}